requirements.txt
streamlit
nbformat
python-dotenv
transformers
torch
numpy
google-generativeai